add pr out and tidy custom webpack plugins
