+++ .prout.json
@@ -1,0 +1,8 @@
+{
+	"checkpoints": {
+		"PROD": {
+			"url": "https://assets-code.guim.co.uk/commercial/prout",
+			"overdue": "10M"
+		}
+	}
+}
